feat: refactor main menu into ui layer
